% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oops.R
\name{Ops.character}
\alias{Ops.character}
\alias{Ops}
\title{Methods sum}
\usage{
\method{Ops}{character}(e1, e2)
}
\arguments{
\item{e1}{character string}

\item{e2}{character string}
}
\description{
Methods sum
}
\examples{



}
