% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringplus_functions.R
\name{set_string_ops}
\alias{set_string_ops}
\title{Set infix string operators}
\usage{
set_string_ops(concat = "&", format = "&", permanent = TRUE)
}
\arguments{
\item{concat}{The concatenation operator (see `string_concat_op`)}

\item{format}{The format operator (see `string_format_op`)}

\item{permanent}{If `TRUE` store these settings to load for future sessions}
}
\description{
Set infix string operators for concatenation and format functions
}
