% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringplus_functions.R
\name{string_concat_op}
\alias{string_concat_op}
\title{Concatenate two strings}
\usage{
string_concat_op(e1, e2)
}
\arguments{
\item{e1}{Must be a character vector}

\item{e2}{Anything, will be the second argument in `paste0`}
}
\value{
If e1 is a character vector, e1 and e2 pasted together. Equivalent to paste0(e1, e2).
}
\description{
Concatenate together two strings. Not intended to be called directly, use `set_string_ops`.
}
\examples{
set_string_ops(concat = "+", permanent = FALSE)
"hello" + "world" # output: "helloworld"
1 + 2 # output: 3
}
