% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringplus_functions.R
\name{string_format_op}
\alias{string_format_op}
\title{Format a string}
\usage{
string_format_op(e1, e2)
}
\arguments{
\item{e1}{Must be a character vector}

\item{e2}{Anything, will be used as subsequent arguments.}
}
\value{
If e1 is a character vector, e1 and e2 will be combined using either `sprintf`
or `glue::glue` depending on whether e2 is named.
}
\description{
Function for formatting a string using `glue::glue` or `sprintf`. Not intended to be called directly, use `set_string_ops`.
}
\examples{
set_string_ops(format = "*", permanent = FALSE)
"C:/folder/\%s/file.txt" * "subfolder"

# Equivalent to 
sprintf("C:/folder/\%s/file.txt", "subfolder")

"C:/folder/{var}/file.txt" * c(var="subfolder")

# Equivalent to 
glue::glue("C:/folder/{var}/file.txt", var="subfolder")

1 * 2 # output: 2
}
