% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{+}
\alias{+}
\title{Sum operations among two or more character string}
\usage{
"+"(e1, e2)
}
\arguments{
\item{e1, e2}{addends}
}
\description{
Sum operations among two or more character string
}
\examples{

subject <- "Susan"
verb <- "is eating"
object <- "an apple"

out <- subject+" "+verb+" "+object

dir <- "~/mydrir/"
files <- c("myfile.txt","readme.txt")

out <- dir+files


}
